+++ scrapy.cfg
@@ -4,8 +4,8 @@
 # https://scrapyd.readthedocs.io/en/latest/deploy.html
 
 [settings]
-default = youlaparse.settings
+default = gbparsers.settings
 
 [deploy]
 #url = http://localhost:6800/
-project = youlaparse
+project = gbparsers
